feat: 🎸 adding csv extration routine

+++ src/test/unit/airport/airport-test.test.js
@@ -1,4 +1,7 @@
-const { registerRoute } = require('../../../modules/airport/airport-service');
+const {
+  registerRoute,
+  extractAirpoirtsFromCsv,
+} = require('../../../modules/airport/airport-service');
 const fs = require('fs');
 const { INPUT_ROUTES_FILE_NAME } = process.env;
 
@@ -8,10 +11,18 @@ describe('Test airport-service functions', () => {
       await registerRoute({ origin: 'XXX', destination: 'YYY', value: 10 })
     ).toEqual('Route XXX to YYY for 10');
   });
+
+  test('Test extractAirpoirtsFromCsv()', async () => {
+    expect(await extractAirpoirtsFromCsv()).toEqual({
+      XXX: [{ YYY: 10 }],
+    });
+  });
 });
 
 afterAll(() => {
-  fs.unlinkSync(
-    `${__dirname}/../../../modules/airport/artifacts/${INPUT_ROUTES_FILE_NAME}`
-  );
+  if (INPUT_ROUTES_FILE_NAME) {
+    fs.unlinkSync(
+      `${__dirname}/../../../modules/airport/artifacts/${INPUT_ROUTES_FILE_NAME}`
+    );
+  }
 });
